# market_go_api_gateway/api/docs/swagger.yaml
definitions:
  branch_service.Branch:
    properties:
      address:
        type: string
      created_at:
        type: string
      deleted_at:
        type: string
      founded_at:
        type: string
      id:
        type: string
      name:
        type: string
      updated_at:
        type: string
    type: object
  branch_service.BranchCreateReq:
    properties:
      address:
        type: string
      founded_at:
        type: string
      name:
        type: string
    type: object
  branch_service.BranchCreateResp:
    properties:
      msg:
        type: string
    type: object
  branch_service.BranchDeleteResp:
    properties:
      msg:
        type: string
    type: object
...
  branch_service.BranchUpdateReq:
    properties:
      address:
        type: string
      founded_at:
        type: string
      id:
        type: string
      name:
        type: string
    type: object
  branch_service.BranchUpdateResp:
    properties:
      msg:
        type: string
    type: object
  handler.Response:
    properties:
      data: {}
      description:
        type: string
      status:
        type: integer
    type: object
...
info:
  contact: {}
paths:
...
  /v1/branches:
    get:
      consumes:
      - application/json
      description: Get branches
      parameters:
      - default: 10
        description: limit for response
        in: query
        name: limit
        type: integer
      - default: 1
        description: page for response
        in: query
        name: page
        type: integer
      - description: search by name and address
        in: query
        name: search
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/branch_service.Branch'
            type: array
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/handler.Response'
            - properties:
                data:
                  type: string
              type: object
        "404":
          description: Not Found
          schema:
            allOf:
            - $ref: '#/definitions/handler.Response'
            - properties:
                data:
                  type: string
              type: object
        "500":
          description: Internal Server Error
          schema:
            allOf:
            - $ref: '#/definitions/handler.Response'
            - properties:
                data:
                  type: string
              type: object
      summary: List branches
      tags:
      - Branches
    post:
      consumes:
      - application/json
      description: Create a new branch with the provided details
      parameters:
      - description: data of the branch
        in: body
        name: branch
        required: true
        schema:
          $ref: '#/definitions/branch_service.BranchCreateReq'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/branch_service.BranchCreateResp'
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/handler.Response'
            - properties:
                data:
                  type: string
              type: object
        "404":
          description: Not Found
          schema:
            allOf:
            - $ref: '#/definitions/handler.Response'
            - properties:
                data:
                  type: string
              type: object
        "500":
          description: Internal Server Error
          schema:
            allOf:
            - $ref: '#/definitions/handler.Response'
            - properties:
                data:
                  type: string
              type: object
      summary: Create a new branch
      tags:
      - Branches
  /v1/branches/{id}:
    delete:
      consumes:
      - application/json
      description: Delete a branch by its unique identifier
      parameters:
      - description: Branch ID to retrieve
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/branch_service.BranchDeleteResp'
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/handler.Response'
            - properties:
                data:
                  type: string
              type: object
        "404":
          description: Not Found
          schema:
            allOf:
            - $ref: '#/definitions/handler.Response'
            - properties:
                data:
                  type: string
              type: object
        "500":
          description: Internal Server Error
          schema:
            allOf:
            - $ref: '#/definitions/handler.Response'
            - properties:
                data:
                  type: string
              type: object
      summary: Delete a branch
      tags:
      - Branches
    get:
      consumes:
      - application/json
      description: Retrieve a branch by its unique identifier
      parameters:
      - description: Branch ID to retrieve
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/branch_service.Branch'
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/handler.Response'
            - properties:
                data:
                  type: string
              type: object
        "404":
          description: Not Found
          schema:
            allOf:
            - $ref: '#/definitions/handler.Response'
            - properties:
                data:
                  type: string
              type: object
        "500":
          description: Internal Server Error
          schema:
            allOf:
            - $ref: '#/definitions/handler.Response'
            - properties:
                data:
                  type: string
              type: object
      summary: Get a branch by ID
      tags:
      - Branches
    put:
      consumes:
      - application/json
      description: Update an existing branch with the provided details
      parameters:
      - description: Branch ID to update
        in: path
        name: id
        required: true
        type: string
      - description: Updated data for the branch
        in: body
        name: branch
        required: true
        schema:
          $ref: '#/definitions/branch_service.BranchUpdateReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/branch_service.BranchUpdateResp'
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/handler.Response'
            - properties:
                data:
                  type: string
              type: object
        "404":
          description: Not Found
          schema:
            allOf:
            - $ref: '#/definitions/handler.Response'
            - properties:
                data:
                  type: string
              type: object
        "500":
          description: Internal Server Error
          schema:
            allOf:
            - $ref: '#/definitions/handler.Response'
            - properties:
                data:
                  type: string
              type: object
      summary: Update an existing branch
      tags:
      - Branches
...
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"